requirements.txt
Pillow
opencv-python
tk